abstract get current focus editor selection text method
