+++ src/vs/workbench/contrib/search/browser/searchActionsTextQuickAccess.ts
@@ -12,9 +12,9 @@ import { IQuickInputService } from 'vs/platform/quickinput/common/quickInput';
 import { TEXT_SEARCH_QUICK_ACCESS_PREFIX } from 'vs/workbench/contrib/search/browser/quickTextSearch/textSearchQuickAccess';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { IEditor } from 'vs/editor/common/editorCommon';
-import { isCodeEditor } from 'vs/editor/browser/editorBrowser';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IEditorOptions } from 'vs/editor/common/config/editorOptions';
+import { getSelectionTextFromEditor } from 'vs/editor/contrib/find/browser/findController';
 
 registerAction2(class TextSearchQuickAccessAction extends Action2 {
 
@@ -34,9 +34,8 @@ registerAction2(class TextSearchQuickAccessAction extends Action2 {
 
 	override async run(accessor: ServicesAccessor, match: RenderableMatch | undefined): Promise<any> {
 		const quickInputService = accessor.get(IQuickInputService);
-		const searchText = getSearchText(accessor);
-		const searchContents = searchText ? ' ' + searchText : '';
-		quickInputService.quickAccess.show(TEXT_SEARCH_QUICK_ACCESS_PREFIX + searchContents);
+		const searchText = getSearchText(accessor) ?? '';
+		quickInputService.quickAccess.show(TEXT_SEARCH_QUICK_ACCESS_PREFIX + searchText);
 	}
 });
 
@@ -47,6 +46,7 @@ function getSearchText(accessor: ServicesAccessor): string | null {
 	if (editorService.activeTextEditorControl === undefined) {
 		return null;
 	}
+
 	const activeEditor: IEditor = editorService.activeTextEditorControl;
 	if (!activeEditor) {
 		return null;
@@ -61,32 +61,5 @@ function getSearchText(accessor: ServicesAccessor): string | null {
 		return null;
 	}
 
-	const range = activeEditor.getSelection();
-	if (!range) {
-		return null;
-	}
-	if (!isCodeEditor(activeEditor) || !activeEditor.hasModel()) {
-		return null;
-	}
-
-	let searchText = '';
-	for (let i = range.startLineNumber; i <= range.endLineNumber; i++) {
-		let lineText = activeEditor.getModel().getLineContent(i);
-		if (i === range.endLineNumber) {
-			lineText = lineText.substring(0, range.endColumn - 1);
-		}
-
-		if (i === range.startLineNumber) {
-			lineText = lineText.substring(range.startColumn - 1);
-		}
-
-		if (i !== range.startLineNumber) {
-			lineText = '\n' + lineText;
-		}
-
-		searchText += lineText;
-	}
-
-	return searchText;
+	return getSelectionTextFromEditor(false, activeEditor, configurationService);
 }
-
